move character implementation to character folder
