+++ lmfdb/characters/HeckeCharacters.py
